Improved commenting on Peplinski soil model.

+++ gprMax/receivers.py
@@ -72,7 +72,7 @@ def gpu_get_rx_array(rxs_gpu, rxcoords_gpu, G):
     """Copy output from receivers array used on GPU back to receiver objects.
 
     Args:
-        rxs_gpu (int): numpy array of receiver data from GPU - rows are field components; columns are iterations; pages are receivers.
+        rxs_gpu (float): numpy array of receiver data from GPU - rows are field components; columns are iterations; pages are receivers.
         rxcoords_gpu (float): numpy array of receiver coordinates from GPU.
         G (class): Grid class instance - holds essential parameters describing the model.
     """
